Consolidate layout code
