+++ src/applications/vaos/components/AppUnavailable.jsx
@@ -4,7 +4,7 @@ import AlertBox from '@department-of-veterans-affairs/formation-react/AlertBox';
 
 export default function AppUnavailable() {
   return (
-    <div className="vads-l-grid-container vads-u-padding-x--2p5 large-screen:vads-u-padding-x--0 vads-u-padding-bottom--2p5">
+    <>
       <Breadcrumbs />
       <div className="vads-l-row">
         <div className="vads-l-col--12 medium-screen:vads-l-col--8 vads-u-margin-bottom--4">
@@ -18,6 +18,6 @@ export default function AppUnavailable() {
           </AlertBox>
         </div>
       </div>
-    </div>
+    </>
   );
 }
