+++ src/applications/vaos/containers/VAOSApp.jsx
@@ -10,98 +10,54 @@ import DowntimeNotification, {
 } from 'platform/monitoring/DowntimeNotification';
 import environment from 'platform/utilities/environment';
 
-import { captureError } from '../utils/error';
 import {
   vaosApplication,
   selectFeatureToggleLoading,
 } from '../utils/selectors';
 import NoRegistrationMessage from '../components/NoRegistrationMessage';
 import AppUnavailable from '../components/AppUnavailable';
-import ErrorMessage from '../components/ErrorMessage';
 import DowntimeMessage from '../components/DowntimeMessage';
+import FullWidthLayout from '../components/FullWidthLayout';
 
-export class VAOSApp extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      hasError: false,
-    };
-  }
+function VAOSApp({
+  user,
+  children,
+  showApplication,
+  loadingFeatureToggles,
+  sites,
+}) {
+  const hasRegisteredSystems = sites?.length > 0;
 
-  static getDerivedStateFromError() {
-    return { hasError: true };
-  }
-
-  componentDidCatch(error) {
-    captureError(error);
-  }
-
-  render() {
-    const {
-      user,
-      children,
-      showApplication,
-      loadingFeatureToggles,
-      sites,
-    } = this.props;
-
-    const hasRegisteredSystems = sites?.length > 0;
-
-    if (this.state.hasError) {
-      return (
-        <div className="vads-l-grid-container vads-u-padding-x--2p5 large-screen:vads-u-padding-x--0 vads-u-padding-bottom--2p5">
-          <div className="vads-l-row">
-            <div className="vads-l-col--12 medium-screen:vads-l-col--8 vads-u-margin-bottom--4">
-              <ErrorMessage />
-            </div>
-          </div>
-        </div>
-      );
-    }
-
-    return (
-      <RequiredLoginView
-        serviceRequired={[
-          backendServices.USER_PROFILE,
-          backendServices.FACILITIES,
-        ]}
-        user={user}
-        verify={!environment.isLocalhost()}
-      >
-        {loadingFeatureToggles && (
-          <div className="vads-l-grid-container vads-u-padding-x--2p5 large-screen:vads-u-padding-x--0 vads-u-padding-bottom--2p5">
-            <div className="vads-l-row">
-              <div className="vads-l-col--12 medium-screen:vads-l-col--8 vads-u-margin-bottom--4">
-                <LoadingIndicator />
-              </div>
-            </div>
-          </div>
+  return (
+    <RequiredLoginView
+      serviceRequired={[
+        backendServices.USER_PROFILE,
+        backendServices.FACILITIES,
+      ]}
+      user={user}
+      verify={!environment.isLocalhost()}
+    >
+      {loadingFeatureToggles && (
+        <FullWidthLayout>
+          <LoadingIndicator />
+        </FullWidthLayout>
+      )}
+      {!loadingFeatureToggles &&
+        showApplication && (
+          <DowntimeNotification
+            appTitle="VA online scheduling tool"
+            dependencies={[externalServices.mvi, externalServices.vaos]}
+            render={(props, childContent) => (
+              <DowntimeMessage {...props}>{childContent}</DowntimeMessage>
+            )}
+          >
+            {!hasRegisteredSystems && <NoRegistrationMessage />}
+            {hasRegisteredSystems && children}
+          </DowntimeNotification>
         )}
-        {!loadingFeatureToggles &&
-          showApplication && (
-            <DowntimeNotification
-              appTitle="VA online scheduling tool"
-              dependencies={[externalServices.mvi, externalServices.vaos]}
-              render={(props, childContent) => (
-                <DowntimeMessage {...props}>{childContent}</DowntimeMessage>
-              )}
-            >
-              {!hasRegisteredSystems && (
-                <div className="vads-l-grid-container vads-u-padding-x--2p5 large-screen:vads-u-padding-x--0 vads-u-padding-bottom--2p5">
-                  <div className="vads-l-row">
-                    <div className="vads-l-col--12 vads-u-margin-bottom--4">
-                      <NoRegistrationMessage />
-                    </div>
-                  </div>
-                </div>
-              )}
-              {hasRegisteredSystems && children}
-            </DowntimeNotification>
-          )}
-        {!loadingFeatureToggles && !showApplication && <AppUnavailable />}
-      </RequiredLoginView>
-    );
-  }
+      {!loadingFeatureToggles && !showApplication && <AppUnavailable />}
+    </RequiredLoginView>
+  );
 }
 
 function mapStateToProps(state) {
